+++ app/ante/utils_test.go
@@ -25,16 +25,20 @@ import (
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/tx"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
+	kmultisig "github.com/cosmos/cosmos-sdk/crypto/keys/multisig"
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
+	"github.com/cosmos/cosmos-sdk/crypto/types/multisig"
 	"github.com/cosmos/cosmos-sdk/simapp"
 	"github.com/cosmos/cosmos-sdk/testutil/testdata"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	txtypes "github.com/cosmos/cosmos-sdk/types/tx"
 	"github.com/cosmos/cosmos-sdk/types/tx/signing"
+	sdkante "github.com/cosmos/cosmos-sdk/x/auth/ante"
 	authsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
 	authtx "github.com/cosmos/cosmos-sdk/x/auth/tx"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	cryptocodec "github.com/evmos/ethermint/crypto/codec"
+	"github.com/evmos/ethermint/crypto/ethsecp256k1"
 
 	"github.com/cosmos/cosmos-sdk/crypto/keys/ed25519"
 	evtypes "github.com/cosmos/cosmos-sdk/x/evidence/types"
@@ -111,6 +115,7 @@ func (suite *AnteTestSuite) SetupTest() {
 	encodingConfig := encoding.MakeConfig(app.ModuleBasics)
 	// We're using TestMsg amino encoding in some tests, so register it here.
 	encodingConfig.Amino.RegisterConcrete(&testdata.TestMsg{}, "testdata.TestMsg", nil)
+	eip712.SetEncodingConfig(encodingConfig)
 
 	suite.clientCtx = client.Context{}.WithTxConfig(encodingConfig.TxConfig)
 
@@ -446,6 +451,183 @@ func (suite *AnteTestSuite) CreateTestEIP712CosmosTxBuilder(
 	suite.Require().NoError(err)
 
 	return builder
+}
+
+// Generate a set of pub/priv keys to be used in creating multi-keys
+func (suite *AnteTestSuite) GenerateMultipleKeys(n int) ([]cryptotypes.PrivKey, []cryptotypes.PubKey) {
+	privKeys := make([]cryptotypes.PrivKey, n)
+	pubKeys := make([]cryptotypes.PubKey, n)
+	for i := 0; i < n; i++ {
+		privKey, err := ethsecp256k1.GenerateKey()
+		suite.Require().NoError(err)
+		privKeys[i] = privKey
+		pubKeys[i] = privKey.PubKey()
+	}
+	return privKeys, pubKeys
+}
+
+// generateSingleSignature signs the given sign doc bytes using the given signType (EIP-712 or Standard)
+func (suite *AnteTestSuite) generateSingleSignature(signMode signing.SignMode, privKey cryptotypes.PrivKey, signDocBytes []byte, signType string) (signature signing.SignatureV2) {
+	var (
+		msg []byte
+		err error
+	)
+
+	msg = signDocBytes
+
+	if signType == "EIP-712" {
+		msg, err = eip712.GetEIP712HashForMsg(signDocBytes)
+		suite.Require().NoError(err)
+	}
+
+	sigBytes, _ := privKey.Sign(msg)
+	sigData := &signing.SingleSignatureData{
+		SignMode:  signMode,
+		Signature: sigBytes,
+	}
+
+	return signing.SignatureV2{
+		PubKey: privKey.PubKey(),
+		Data:   sigData,
+	}
+}
+
+// generateMultikeySignatures signs a set of messages using each private key within a given multi-key
+func (suite *AnteTestSuite) generateMultikeySignatures(signMode signing.SignMode, privKeys []cryptotypes.PrivKey, signDocBytes []byte, signType string) (signatures []signing.SignatureV2) {
+	n := len(privKeys)
+	signatures = make([]signing.SignatureV2, n)
+
+	for i := 0; i < n; i++ {
+		privKey := privKeys[i]
+		currentType := signType
+
+		// If mixed type, alternate signing type on each iteration
+		if signType == "mixed" {
+			if i%2 == 0 {
+				currentType = "EIP-712"
+			} else {
+				currentType = "Standard"
+			}
+		}
+
+		signatures[i] = suite.generateSingleSignature(
+			signMode,
+			privKey,
+			signDocBytes,
+			currentType,
+		)
+	}
+
+	return signatures
+}
+
+// RegisterAccount creates an account with the keeper and populates the initial balance
+func (suite *AnteTestSuite) RegisterAccount(pubKey cryptotypes.PubKey, balance *big.Int) {
+	acc := suite.app.AccountKeeper.NewAccountWithAddress(suite.ctx, sdk.AccAddress(pubKey.Address()))
+	suite.app.AccountKeeper.SetAccount(suite.ctx, acc)
+
+	suite.app.EvmKeeper.SetBalance(suite.ctx, common.BytesToAddress(pubKey.Address()), balance)
+}
+
+// createSignerBytes generates sign doc bytes using the given parameters
+func (suite *AnteTestSuite) createSignerBytes(chainId string, signMode signing.SignMode, pubKey cryptotypes.PubKey, txBuilder client.TxBuilder) []byte {
+	acc, err := sdkante.GetSignerAcc(suite.ctx, suite.app.AccountKeeper, sdk.AccAddress(pubKey.Address()))
+	suite.Require().NoError(err)
+	signerInfo := authsigning.SignerData{
+		Address:       sdk.MustBech32ifyAddressBytes(sdk.GetConfig().GetBech32AccountAddrPrefix(), acc.GetAddress().Bytes()),
+		ChainID:       chainId,
+		AccountNumber: acc.GetAccountNumber(),
+		Sequence:      acc.GetSequence(),
+		PubKey:        pubKey,
+	}
+
+	signerBytes, err := suite.clientCtx.TxConfig.SignModeHandler().GetSignBytes(
+		signMode,
+		signerInfo,
+		txBuilder.GetTx(),
+	)
+	suite.Require().NoError(err)
+
+	return signerBytes
+}
+
+// createBaseTxBuilder creates a TxBuilder to be used for Single- or Multi-signing
+func (suite *AnteTestSuite) createBaseTxBuilder(msg sdk.Msg, gas uint64) client.TxBuilder {
+	txBuilder := suite.clientCtx.TxConfig.NewTxBuilder()
+
+	txBuilder.SetGasLimit(gas)
+	txBuilder.SetFeeAmount(sdk.NewCoins(
+		sdk.NewCoin("aphoton", sdk.NewInt(10000)),
+	))
+
+	err := txBuilder.SetMsgs(msg)
+	suite.Require().NoError(err)
+
+	txBuilder.SetMemo("")
+
+	return txBuilder
+}
+
+// CreateTestSignedMultisigTx creates and sign a multi-signed tx for the given message. `signType` indicates whether to use standard signing ("Standard"),
+// EIP-712 signing ("EIP-712"), or a mix of the two ("mixed").
+func (suite *AnteTestSuite) CreateTestSignedMultisigTx(privKeys []cryptotypes.PrivKey, signMode signing.SignMode, msg sdk.Msg, chainId string, gas uint64, signType string) client.TxBuilder {
+	pubKeys := make([]cryptotypes.PubKey, len(privKeys))
+	for i, privKey := range privKeys {
+		pubKeys[i] = privKey.PubKey()
+	}
+
+	// Re-derive multikey
+	numKeys := len(privKeys)
+	multiKey := kmultisig.NewLegacyAminoPubKey(numKeys, pubKeys)
+
+	suite.RegisterAccount(multiKey, big.NewInt(10000000000))
+
+	txBuilder := suite.createBaseTxBuilder(msg, gas)
+
+	// Prepare signature field
+	sig := multisig.NewMultisig(len(pubKeys))
+	txBuilder.SetSignatures(signing.SignatureV2{
+		PubKey: multiKey,
+		Data:   sig,
+	})
+
+	signerBytes := suite.createSignerBytes(chainId, signMode, multiKey, txBuilder)
+
+	// Sign for each key and update signature field
+	sigs := suite.generateMultikeySignatures(signMode, privKeys, signerBytes, signType)
+	for _, pkSig := range sigs {
+		err := multisig.AddSignatureV2(sig, pkSig, pubKeys)
+		suite.Require().NoError(err)
+	}
+
+	txBuilder.SetSignatures(signing.SignatureV2{
+		PubKey: multiKey,
+		Data:   sig,
+	})
+
+	return txBuilder
+}
+
+func (suite *AnteTestSuite) CreateTestSingleSignedTx(privKey cryptotypes.PrivKey, signMode signing.SignMode, msg sdk.Msg, chainId string, gas uint64, signType string) client.TxBuilder {
+	pubKey := privKey.PubKey()
+
+	suite.RegisterAccount(pubKey, big.NewInt(10000000000))
+
+	txBuilder := suite.createBaseTxBuilder(msg, gas)
+
+	// Prepare signature field
+	sig := signing.SingleSignatureData{}
+	txBuilder.SetSignatures(signing.SignatureV2{
+		PubKey: pubKey,
+		Data:   &sig,
+	})
+
+	signerBytes := suite.createSignerBytes(chainId, signMode, pubKey, txBuilder)
+
+	sigData := suite.generateSingleSignature(signMode, privKey, signerBytes, signType)
+	txBuilder.SetSignatures(sigData)
+
+	return txBuilder
 }
 
 func NextFn(ctx sdk.Context, _ sdk.Tx, _ bool) (sdk.Context, error) {